requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-generativeai==0.5.2
python-dotenv==1.0.1
langgraph==0.0.28
httpx==0.27.0
typing-extensions==4.12.1
